refactor(w3c): separate scenarios in w3c
Left w3c scenario that includes all scenarios for back-compatibility and usability
